abstract downloadtemplate, renameservice functions
