+++ lib/utils/renameService.test.js
@@ -1,0 +1,108 @@
+'use strict';
+
+const expect = require('chai').expect;
+const Serverless = require('../Serverless');
+const testUtils = require('../../tests/utils');
+const fse = require('fs-extra');
+const path = require('path');
+
+const { renameService } = require('./renameService');
+
+describe('renameService', () => {
+  let serverless;
+  let cwd;
+
+  let servicePath;
+  let newServicePath;
+
+  beforeEach(() => {
+    const tmpDir = testUtils.getTmpDirPath();
+    cwd = process.cwd();
+
+    fse.mkdirsSync(tmpDir);
+    process.chdir(tmpDir);
+
+    servicePath = tmpDir;
+    newServicePath = path.join(servicePath, 'new-service-name');
+
+    serverless = new Serverless();
+    serverless.init();
+  });
+
+  afterEach(() => {
+    // change back to the old cwd
+    process.chdir(cwd);
+  });
+
+  it('should set new service in serverless.yml and name in package.json', () => {
+    const defaultServiceYml =
+      'service: service-name\n\nprovider:\n  name: aws\n';
+    const newServiceYml =
+      'service: new-service-name\n\nprovider:\n  name: aws\n';
+
+    const defaultServiceName = 'service-name';
+    const newServiceName = 'new-service-name';
+
+    const packageFile = path.join(servicePath, 'package.json');
+    const serviceFile = path.join(servicePath, 'serverless.yml');
+
+    serverless.utils.writeFileSync(packageFile, { name: defaultServiceName });
+    fse.writeFileSync(serviceFile, defaultServiceYml);
+
+    renameService(newServiceName, servicePath);
+    const serviceYml = fse.readFileSync(serviceFile, 'utf-8');
+    const packageJson = serverless.utils.readFileSync(packageFile);
+    expect(serviceYml).to.equal(newServiceYml);
+    expect(packageJson.name).to.equal(newServiceName);
+  });
+
+  it('should set new service in commented serverless.yml and name in package.json', () => {
+    const defaultServiceYml =
+      '# comment\nservice: service-name #comment\n\nprovider:\n  name: aws\n# comment';
+    const newServiceYml =
+      '# comment\nservice: new-service-name #comment\n\nprovider:\n  name: aws\n# comment';
+
+    const defaultServiceName = 'service-name';
+    const newServiceName = 'new-service-name';
+
+    const packageFile = path.join(servicePath, 'package.json');
+    const serviceFile = path.join(servicePath, 'serverless.yml');
+
+    serverless.utils.writeFileSync(packageFile, { name: defaultServiceName });
+    fse.writeFileSync(serviceFile, defaultServiceYml);
+
+    renameService(newServiceName, servicePath);
+    const serviceYml = fse.readFileSync(serviceFile, 'utf-8');
+    const packageJson = serverless.utils.readFileSync(packageFile);
+    expect(serviceYml).to.equal(newServiceYml);
+    expect(packageJson.name).to.equal(newServiceName);
+  });
+
+  it('should set new service in commented serverless.yml without existing package.json', () => {
+    const defaultServiceYml =
+      '# comment\nservice: service-name #comment\n\nprovider:\n  name: aws\n# comment';
+    const newServiceYml =
+      '# comment\nservice: new-service-name #comment\n\nprovider:\n  name: aws\n# comment';
+
+    const serviceFile = path.join(servicePath, 'serverless.yml');
+
+    serverless.utils.writeFileDir(serviceFile);
+    fse.writeFileSync(serviceFile, defaultServiceYml);
+
+    renameService('new-service-name', servicePath);
+    const serviceYml = fse.readFileSync(serviceFile, 'utf-8');
+    expect(serviceYml).to.equal(newServiceYml);
+  });
+
+  it('should fail to set new service name in serverless.yml', () => {
+    const defaultServiceYml =
+      '# comment\nservice: service-name #comment\n\nprovider:\n  name: aws\n# comment';
+
+    const serviceFile = path.join(servicePath, 'serverledss.yml');
+
+    serverless.utils.writeFileDir(serviceFile);
+    fse.writeFileSync(serviceFile, defaultServiceYml);
+
+    expect(() => renameService('new-service-name', servicePath)).to.throw(Error);
+  });
+});
